Reorganize some modules
