+++ virtool/db/hmm.py
@@ -1,9 +1,11 @@
+import asyncio
 import json
 import logging
 import os
 import shutil
 
 import aiofiles
+import semver
 
 import virtool.db.processes
 import virtool.db.status
@@ -189,3 +191,73 @@ async def purge(db):
             "hidden": True
         }
     })
+
+
+async def get_hmm_status(db):
+    status = await db.status.find_one("hmm")
+
+    status = virtool.utils.base_processor(status)
+
+    status["updating"] = len(status["updates"]) > 1 and status["updates"][-1]["ready"]
+
+    del status["updates"]
+
+    return status
+
+
+async def fetch_and_update_hmm_release(app):
+    """
+    Return the HMM install status document or create one if none exists.
+
+    :param app: the app object
+    :type app: :class:`aiohttp.web.Application`
+
+    """
+    db = app["db"]
+    settings = app["settings"]
+    session = app["client"]
+
+    etag = None
+
+    document = await db.status.find_one("hmm", ["release", "updates"])
+
+    existing = document.get("release", None)
+
+    try:
+        installed = document["updates"][-1]
+    except (IndexError, KeyError):
+        installed = None
+
+    if existing:
+        etag = existing.get("etag", None)
+
+    release = await virtool.github.get_release(settings, session, "virtool/virtool-hmm", etag)
+
+    if release:
+        release = virtool.github.format_release(release)
+    else:
+        release = existing
+
+    release["newer"] = bool(
+        release is None or (
+            installed and
+            semver.compare(release["name"].lstrip("v"), installed["name"].lstrip("v")) == 1
+        )
+    )
+
+    await db.status.update_one({"_id": "hmm"}, {
+        "$set": {
+            "release": release
+        }
+    }, upsert=True)
+
+    return release
+
+
+async def refresh(app):
+    try:
+        while True:
+            await fetch_and_update_hmm_release(app)
+            await asyncio.sleep(600, loop=app.loop)
+    except asyncio.CancelledError:
+        pass
